Enable Redis Queues

+++ lumen/app/Jobs/ImportEvent.php
@@ -36,8 +36,8 @@ class ImportEvent extends Job
 
         if (empty($event)) {
             $venueStorage = $client->test->venues;
-            $venueId = data_get($toImport, 'metadata.venue_id');
-            
+            $venueId = (int) data_get($toImport, 'metadata.venue_id');
+
             $venue = $venueStorage->findOne([
                 'external_id' => $venueId
             ]);
@@ -47,7 +47,6 @@ class ImportEvent extends Job
                 return $eventStorage->insertOne($toImport);
             }
 
-            // TODO: turn this to Queues
             dispatch(new ImportVenue($venueId));
             dispatch(new ImportEvent($toImport));
         }
